requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
email-validator==2.1.0
python-dotenv==1.0.0